Explain error of Invoke-WebRequest failing:
`ERROR: The response content cannot be parsed because the Internet Explorer engine is not available`

This requires Internet Explorer to be run at least once by the elevated
user.

Right now this might block moderator approval for this package (it
should have blocked in the past) as `Invoke-WebRequest` requires
PowerShell version 3, but chocolatey limits to PowerShell version 2.